Drop GCC 2.7.2: the game was built entirely with 2.6.3
All eight objects that were pinned to 2.7.2 now match under 2.6.3, so
the second compiler is gone from the build.

One mechanism explains most of them. 2.7.2 is the less aggressive
compiler on narrowing casts: it keeps the andi 0xff and andi 0xffff that
2.6.3's nonzero_bits proves redundant and drops. Any decompilation that
writes a redundant cast where the original had a narrow type will
therefore look like it needs 2.7.2. Four of the eight were exactly that,
and the fix is to give the variable the narrow type instead of masking
an s32.

track/GameLoadEnvironmentCue was pinned only because cc1 2.6.3 segfaults
on __attribute__((packed)) applied to a struct. It accepts the attribute
on a field, and then produces identical text and emits lwl/lwr for the
unaligned load where 2.7.2 assembles the word byte by byte. That is a
compiler crash, not a codegen requirement.

Two were statement placement: a temporary holding g_FadeStep flipped the
operand order of an addu, and a pointer increment sitting mid loop made
2.6.3 duplicate a branch target instead of filling a delay slot.

Two needed no source change at all; they had been pinned on a
measurement that compared disassembly text, where LM debug label numbers
read as instruction differences. Compare .text bytes.

Nothing distinguished the Sony library units in lib/libsnd and lib/libspu
from Namco's own code: they fell to the same class of fix, which rules
out the theory that the libraries were shipped prebuilt from a different
SDK release.

_SpuSetAnyVoice also loses eleven lines of register pinning, which had
been blocking the cross jump that merges its two stores.

The 2.7.2 branch in tools/scripts/cc.sh and its download logic are now
unreachable from the build.

+++ src/main/PAL/lib/libsnd/SsSeqOpen.c
@@ -10,7 +10,7 @@ s32 SsSeqOpen(s32 seq_data, s32 vab_id) asm("func_8006F004");
 s32 SsSeqOpen(s32 seq_data, s32 vab_id) {
     s32 used;
     s32 i;
-    s32 found;
+    u8 found;
     s32 slot;
 
     used = D_801E4B80;
@@ -27,7 +27,7 @@ s32 SsSeqOpen(s32 seq_data, s32 vab_id) {
             found = 1;
         }
         i++;
-    } while ((u8)found == 0);
+    } while (found == 0);
 
     D_801E4B80 = (1 << (s16)slot) | D_801E4B80;
